Move SimulateCallOutcome / SimulateOutcome out of Wallet.java into jsonrpc/types/
